requirements.txt
paho-mqtt
psutil
"fastapi[standard]"